Replace ResultsPage with new version
