+++ frontend/src/pages/ResultsPage.tsx
@@ -1,93 +1,114 @@
-import React, { useEffect, useState } from "react";
-import { Table } from "@equinor/eds-core-react";
-import { getLabResults, concentrations } from "../api/api";
-import { ExperimentResult } from "../dto/ExperimentResult";
+import React, { useState } from "react";
+import { EdsDataGrid } from "@equinor/eds-data-grid-react";
+import { getLabResults } from "../api/api";
+import { useQuery } from "@tanstack/react-query";
 
 const ResultsPage: React.FC = () => {
-    const [results, setResults] = useState<ExperimentResult[]>([]);
-    const [loading, setLoading] = useState(true);
-    const [error, setErrorr] = useState<string | null>(null);
-    const [headers, setHeaders] = useState<Record<string, string[]>>({});
-    //const { instance, accounts } = useMsal();
-    //const account = useAccount(accounts[0] || {});
+    const initialPrefix = "in-";
+    const finalPrefix = "out-";
+    const [enableFilters, setEnableFilters] = useState(false);
+    const {
+        data: labResults,
+        error,
+        isLoading,
+    } = useQuery({
+        queryKey: ["results"],
+        queryFn: () => getLabResults(),
+    });
 
-    useEffect(() => {
-        const fetchData = async () => {
-            try {
-                const data: ExperimentResult[] = await getLabResults();
-                setResults(data);
-                const initial_conc = Array.from(
-                    new Set(data.flatMap((entry) => [...Object.keys(entry.initial_concentrations)]))
-                );
-                const final_conc = Array.from(
-                    new Set(data.flatMap((entry) => [...Object.keys(entry.final_concentrations)]))
-                );
-                setHeaders({ initial_concentrations: initial_conc, final_concentrations: final_conc });
-            } catch (error) {
-                setErrorr(String(error));
-            } finally {
-                setLoading(false);
-            }
-        };
-
-        fetchData();
-    }, []);
-
-    if (loading) {
-        return <div>Loading...</div>;
+    if (isLoading) {
+        return <>Fetching results ...</>;
     }
-
     if (error) {
-        return <div>Error: {error}</div>;
+        return <>Could not fetch results.</>;
     }
+
+    const initialConcHeaders = Array.from(
+        new Set(labResults!.flatMap((entry) => [...Object.keys(entry.initial_concentrations)]))
+    );
+    const finalConcHeaders = Array.from(
+        new Set(labResults!.flatMap((entry) => [...Object.keys(entry.final_concentrations)]))
+    );
+
+    const columns = [
+        {
+            columns: [
+                {
+                    header: "Experiment",
+                    id: "name",
+                    accessorKey: "name",
+                    size: 100,
+                },
+                {
+                    header: "Time",
+                    id: "time",
+                    accessorKey: "time",
+                    size: 65,
+                },
+            ],
+            header: "Meta data",
+        },
+        {
+            columns: initialConcHeaders.map((header) => ({
+                accessorKey: initialPrefix + header,
+                header: header,
+                id: initialPrefix + header,
+                size: 65,
+            })),
+            header: "Input Concentrations",
+        },
+        {
+            columns: finalConcHeaders.map((header) => ({
+                accessorKey: finalPrefix + header,
+                header: header,
+                id: finalPrefix + header,
+                size: 65,
+            })),
+            header: "Output Concentrations",
+        },
+    ];
+
+    const rows = labResults!.map((entry) => ({
+        id: entry.name,
+        name: entry.name,
+        time: String(entry.time),
+        ...Object.fromEntries(
+            Object.entries(entry.initial_concentrations).map(([key, value]) => [
+                initialPrefix + key,
+                +Number(value).toPrecision(3),
+            ])
+        ),
+        ...Object.fromEntries(
+            Object.entries(entry.final_concentrations).map(([key, value]) => [
+                finalPrefix + key,
+                +Number(value).toPrecision(3),
+            ])
+        ),
+    }));
+
+    const handleEnableFilters = () => {
+        setEnableFilters(!enableFilters);
+    };
 
     return (
-        <div>
+        <>
             <h1>Results</h1>
-            <Table>
-                <Table.Head>
-                    <Table.Row>
-                        <Table.Cell rowSpan={2}>Experiment</Table.Cell>
-                        <Table.Cell rowSpan={2}>Time</Table.Cell>
-                        <Table.Cell colSpan={Object.keys(headers["initial_concentrations"]).length}>
-                            Input Concentrations
-                        </Table.Cell>
-                        <Table.Cell colSpan={Object.keys(headers["final_concentrations"]).length}>
-                            Output Concentrations
-                        </Table.Cell>
-                    </Table.Row>
-                    <Table.Row>
-                        {headers["initial_concentrations"].map((header) => (
-                            <Table.Cell key={"initial-" + header}>{header}</Table.Cell>
-                        ))}
-                        {headers["final_concentrations"].map((header) => (
-                            <Table.Cell key={"final-" + header}>{header}</Table.Cell>
-                        ))}
-                    </Table.Row>
-                </Table.Head>
-                <Table.Body>
-                    {results.map((dataItem, index) => (
-                        <React.Fragment key={index}>
-                            <Table.Row>
-                                <Table.Cell>{dataItem.name}</Table.Cell>
-                                <Table.Cell>{dataItem.time}</Table.Cell>
-                                {Object.entries(dataItem.initial_concentrations as concentrations).map(
-                                    ([_, value], idx) => (
-                                        <Table.Cell key={idx}>{value}</Table.Cell>
-                                    )
-                                )}
-                                {Object.entries(dataItem.final_concentrations as concentrations).map(
-                                    ([_, value], idx) => (
-                                        <Table.Cell key={idx}>{value}</Table.Cell>
-                                    )
-                                )}
-                            </Table.Row>
-                        </React.Fragment>
-                    ))}
-                </Table.Body>
-            </Table>
-        </div>
-    );
+            <label style={{ display: 'flex', alignItems: 'center' }}>
+                <input
+                type="checkbox"
+                checked={enableFilters}
+                onChange={handleEnableFilters}
+                style={{
+                    transform: 'scale(1.5)',
+                    marginBottom: '20px'
+                }}
+                />
+                <span style={{ fontSize: '18px', marginLeft: '8px' , marginBottom: "17px"}}>
+                    {"Enable filters"}
+                </span>
+            </label>            
+            <EdsDataGrid columns={columns} rows={rows} enableColumnFiltering={enableFilters} />;
+        </>
+    )
 };
-
 export default ResultsPage;
